Move Client properties to parent class, remove unnecessary proerty setters and getters
